requirements.txt
flask==3.0.3
pillow==10.4.0
transformers==4.44.2
torch==2.4.1
nltk==3.9.1